src/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.55.3
jinja2==3.1.4
pydantic==2.10.3
tiktoken==0.8.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
